+++ contrib/fmm_tomography/__init__.py
@@ -4,12 +4,12 @@ It's NOT recommended to put extra code inside this file. Code inside this file w
 executed when this submodule is imported, so adding things in this file can slow down
 the importing process of `espresso`.
 
-For contributors, add any intialisation code for your problem into great_cirle_tracing.py, 
-under the method `__init__()` of the class `GreatCirleTracing`.
+For contributors, add any intialisation code for your problem into fmm_tomography.py, 
+under the method `__init__()` of the class `FmmTomography`.
 
 Don't touch this file unless you know what you are doing :)
 """
 
-from .great_circle_tracing import *
+from .fmm_tomography import *
 
-__all__ = ["GreatCircleTracing"]
+__all__ = [ "FmmTomography" ]