+++ contrib/slug_test/slug_test.py
@@ -114,26 +114,30 @@ class SlugTest(EspressoProblem):
         raise NotImplementedError               # optional
 
     def plot_model(self, model):
-        plt.errorbar(self._xp, self.forward(model), yerr=self._sigma, fmt='.', 
+        fig, ax = plt.subplots()
+        ax.errorbar(self._xp, self.forward(model), yerr=self._sigma, fmt='.', 
                      color="lightcoral", ecolor='lightgrey', ms=10)
-        plt.xscale("log")
-        plt.grid(True, which="both")
-        plt.xlabel('Time elapsed')
-        plt.ylabel('Displacement')
+        ax.set_xscale("log")
+        ax.grid(True, which="both")
+        ax.set_xlabel('Time elapsed')
+        ax.set_ylabel('Displacement')
+        return ax
     
     def plot_data(self, data):
-        plt.errorbar(self._xp, data, yerr=self._sigma, fmt='.', 
+        fig, ax = plt.subplots()
+        ax.errorbar(self._xp, data, yerr=self._sigma, fmt='.', 
                      color="lightcoral", ecolor='lightgrey', ms=10)
-        plt.xscale("log")
-        plt.grid(True, which="both")
-        plt.xlabel('Time elapsed')
-        plt.ylabel('Displacement')
+        ax.set_xscale("log")
+        ax.grid(True, which="both")
+        ax.set_xlabel('Time elapsed')
+        ax.set_ylabel('Displacement')
+        return ax
 
-    def misfit(self, data, data2):              # optional
+    def misfit(self, data1, data2):              # optional
         raise NotImplementedError
 
-    def log_likelihood(self, data, data2): 
-        residual = data - data2
+    def log_likelihood(self, data1, data2): 
+        residual = data1 - data2
         return (-0.5*residual@self.inverse_covariance_matrix@residual.T).item()
     
     def log_prior(self, model):