+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## v0.3.7 (07/06/2023)
+
+### New examples
+
+- `SurfaceWaveTomography` examples
+
+### Changed examples
+
+- Naming changes
+    - `FmWavefrontTracker` -> `FmmTomography`
+    - `XrayTracer` -> `XrayTomography`
+    - `ReceiverFunction` -> `ReceiverFunctionInversion`
+    - `GravityDensity` -> `GravityInversion`
+    - `GreatCircleTracing` -> `SurfaceWaveTomography`
+
+### Espresso core
+
+- Template: Minor naming changes in templates
+- Template: Return `matplotlib.axes.Axes` instead of `matplotlib.figures.Figure` in plotting methods
+
 ## v0.3.6 (19/05/2023)
 
 ### Changed examples